Added 1 AWS answer
The correct answer is:
"There is nothing wrong with this security group rule. Assuming that sg-269afc5e is applied to other resources that are properly secured, this rule allows all traffic to pass through that is also assigned security group sg-269afc5e."

This security group only allows connections from another AWS resources which are also in the same security group.